backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
openai
nltk
rapidfuzz
requests
python-multipart
langgraph